app/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
django-filter==22.1
drf-yasg==1.17.1